+++ CHANGELOG.md
@@ -6,6 +6,18 @@ Written by fffadedev.
 
 ## [Unreleased]
 
+### Added
+
+- Sprite for paddle shrink powerup
+- Forcefield powerup and system
+
+## [v1.2.0] Pre-alpha, 2023-11-29
+
+### Added
+
+- New reversal powerup - reverses direction of ball
+- New sprites for paddle and ball
+
 ## [v1.1.0] Pre-alpha, 2023-11-29
 
 ### Added
@@ -21,3 +33,4 @@ Written by fffadedev.
 - [Back to top](#changelog)
 - [unreleased](#unreleased)
 - [1.1.0](#v1.1.0)
+- [1.2.0](#v1.2.0)